[fix/#94] 클라이언트 패키징 분리
