+++ src/main/java/com/moplus/moplus_server/client/submit/domain/ChildProblemSubmitStatus.java
@@ -8,14 +8,11 @@ public enum ChildProblemSubmitStatus {
     public static ChildProblemSubmitStatus determineStatus(ChildProblemSubmitStatus currentStatus, String memberAnswer, String childProblemAnswer) {
         boolean isCorrect = childProblemAnswer.trim().equals(memberAnswer.trim());
 
-        if (currentStatus == CORRECT) {
-            return isCorrect ? CORRECT : INCORRECT;
-        } else if (currentStatus == INCORRECT) {
-            return isCorrect ? RETRY_CORRECT : INCORRECT;
-        } else if (currentStatus == RETRY_CORRECT) {
-            return isCorrect ? RETRY_CORRECT : INCORRECT;
-        } else {
-            return isCorrect ? CORRECT : INCORRECT;
-        }
+        return switch (currentStatus) {
+            case CORRECT -> isCorrect ? CORRECT : INCORRECT;
+            case INCORRECT -> isCorrect ? RETRY_CORRECT : INCORRECT;
+            case RETRY_CORRECT -> isCorrect ? RETRY_CORRECT : INCORRECT;
+            default -> isCorrect ? CORRECT : INCORRECT;
+        };
     }
 }
